+++ tnlcm/auth/__init__.py
@@ -1,0 +1,4 @@
+from tnlcm.models import UserModel
+
+def get_current_user_from_jwt(jwt_identity):
+    return UserModel.objects(username=jwt_identity).first()