+++ tnlcm/database/database.py
@@ -1,0 +1,7 @@
+from mongoengine import connect
+
+def init_db(app):
+    connect(
+        alias="tnlcm-database-alias",
+        host=app.config["MONGO_URI"]
+    )