+++ tnlcm/routes/__init__.py
@@ -1,0 +1,5 @@
+from .callback import callback_namespace
+from .sixglibrary import sixglibrary_namespace
+from .trial_network import trial_network_namespace
+from .user import user_namespace
+from .verification_token import verification_token_namespace