+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Detailed documentation including different sections about TNLCM.
 - Routes for callback, sixglibrary, trial networks, users and verification.
 - MongoDB database to manage the TNs.
+- Flask-mongoengine as Mongo ORM.
 - Docker compose to create MongoDB database.
 - Integration with [6G-Library](https://github.com/6G-SANDBOX/6G-Library).
 - Connection with Jenkins for the deployment of different components. Currently 8 types of components are running: tn_vxlan, tn_bastion, vm_kvm, vxlan, k8s, open5gs, UERANSIM-gNB and UERANSIM-UE.